+++ packages/observables/src/dom/window.test.js
@@ -2,7 +2,7 @@ import { tap } from 'rxjs';
 import { TestScheduler } from 'rxjs/testing';
 import { beforeEach, test, expect, describe, afterEach } from 'vitest';
 
-import { mockOffline, mockOnline, mockReset } from '../../../test-utils/network.js';
+import { mockOffline, mockOnline, mockReset } from '../../../mock/network.js';
 import { connectionObservable } from './window.js';
 
 // HINT: https://betterprogramming.pub/rxjs-testing-write-unit-tests-for-observables-603af959e251