+++ packages/operators/src/request/cache.js
@@ -1,0 +1,12 @@
+import { ReplaySubject, share, timer } from 'rxjs';
+
+export const cache = ttl => {
+  return source =>
+    source.pipe(
+      share({
+        connector: () => new ReplaySubject(),
+        resetOnComplete: () => timer(ttl),
+        resetOnRefCountZero: () => timer(ttl)
+      })
+    );
+};