+++ packages/operators/src/request/retry.js
@@ -1,11 +1,12 @@
 import {
-  catchError,
   combineLatest,
   concatMap,
   delay,
   filter,
   map,
+  merge,
   of,
+  partition,
   retry,
   tap,
   throwError
@@ -20,12 +21,17 @@ export const networkRetry = ({ timeout = defaultTimeout, count } = {}) => {
 
   return source => {
     return source.pipe(
-      concatMap(resp => (!resp.ok && throwError(() => new Error('invalid request'))) || of(resp)),
+      concatMap(resp => {
+        const [success, error] = partition(of(resp), resp => resp.ok);
+        return merge(
+          success,
+          error.pipe(concatMap(() => throwError(() => new Error('invalid request'))))
+        );
+      }),
       retry({
         count,
         delay: () => determineDelayWhenOnline(timeout, ++counter)
-      }),
-      catchError(e => console.error(e))
+      })
     );
   };
 };
@@ -38,7 +44,7 @@ const determineDelayWhenOnline = (timeout, counter) => {
     tap(valid => (counter = counter * valid)),
     // continue only if all observables are valid
     filter(valid => valid),
-    tap(() => console.log(`retry: request - next: ${counter} in ${timeout(counter)}s`)),
-    delay(timeout(counter) || timeout)
+    tap(() => console.log(`retry: request - next: ${counter} in ${timeout(counter)}ms`)),
+    delay(timeout(counter))
   );
 };